optimize: move service outside of proc_service
