+++ src/scene_server/proc_server/service/init_func.go
