+++ src/scene_server/proc_server/service/procdetail.go
